Fixed issue

+++ Lection 3/task 3.1/ReportApp/Models/Data/Item.cs
@@ -4,9 +4,9 @@ public class Item
 {
     public string Name { get; set; }
 
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
 
-    public int Cost { get; set; }
+    public decimal Cost { get; set; }
 
     public string? Notes { get; set; }
 }
